Adapt to move of AtomicIndirector
AtomicIndirector was moved to score/concurrency.

Adapt unrestricted code to new location to reduce technical debt.
