+++ score/language/safecpp/safe_atomics/BUILD
@@ -24,8 +24,8 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         ":error",
+        "@score_baselibs//score/concurrency:atomic_indirector",
         "@score_baselibs//score/language/safecpp/safe_math/details/addition_subtraction",
-        "@score_baselibs//score/memory/shared:atomic_indirector",
         "@score_baselibs//score/result",
     ],
 )
@@ -46,8 +46,8 @@ cc_gtest_unit_test(
     features = COMPILER_WARNING_FEATURES,
     deps = [
         ":try_atomic_add",
+        "@score_baselibs//score/concurrency:atomic_indirector_mock_binding",
         "@score_baselibs//score/language/safecpp/safe_math/details:test_type_collection",
-        "@score_baselibs//score/memory/shared:atomic_indirector_mock_binding",
     ],
 )
