+++ score/utils/meyer_singleton/meyer_singleton_test.cpp
@@ -12,8 +12,8 @@
  ********************************************************************************/
 #include "score/utils/meyer_singleton/meyer_singleton.h"
 
-#include "score/memory/shared/atomic_indirector.h"
-#include "score/memory/shared/atomic_mock.h"
+#include "score/concurrency/atomic_indirector.h"
+#include "score/concurrency/atomic_mock.h"
 #include "score/utils/meyer_singleton/test/single_test_per_process_fixture.h"
 
 #include <score/assert_support.hpp>
@@ -32,7 +32,7 @@ namespace
 using namespace ::testing;
 
 template <typename T>
-using AtomicIndirectorMock = memory::shared::AtomicIndirectorMock<T>;
+using AtomicIndirectorMock = concurrency::AtomicIndirectorMock<T>;
 
 constexpr int kDummyInt{10U};
 constexpr double kDummyDouble{2.0};
@@ -90,7 +90,7 @@ class AtomicMockGuard
     AtomicMockGuard(AtomicMockGuard&&) noexcept = delete;
     AtomicMockGuard& operator=(AtomicMockGuard&&) & noexcept = delete;
 
-    memory::shared::AtomicMock<InitializationState> atomic_mock_{};
+    concurrency::AtomicMock<InitializationState> atomic_mock_{};
 };
 
 class MeyerSingletonFixture : public test::SingleTestPerProcessFixture
